bitsperity-homegrow/mqtt-bridge/requirements.txt
paho-mqtt==2.1.0
pymongo==4.3.3
python-dotenv==1.0.0
colorlog==6.7.0 